added controllers; edited some repositories and models

+++ Backend/src/main/java/com/teambanana/mealsonwheels/model/Donation.java
@@ -2,6 +2,8 @@ package com.teambanana.mealsonwheels.model;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import com.teambanana.mealsonwheels.Enum.*;
 
@@ -15,7 +17,7 @@ public class Donation {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double amount;
+    private BigDecimal amount;
 
     @Enumerated(EnumType.STRING)
     private DonationMethod method;
